fix test for normal variational distribution; fix #59
